+++ src/libAtomVM/stacktrace.h
@@ -29,9 +29,9 @@
 extern "C" {
 #endif
 
-term stacktrace_create_raw(Context *ctx, Module *mod, int current_offset);
+term stacktrace_create_raw(Context *ctx, Module *mod, size_t current_offset);
 
-term stacktrace_create_raw_mfa(Context *ctx, Module *mod, int current_offset, term module_atom,
+term stacktrace_create_raw_mfa(Context *ctx, Module *mod, size_t current_offset, term module_atom,
     term function_atom, int arity);
 
 /**
